refactor(habitat): classify hook model by kind
