+++ tools/habitat-harness/src/service/model/hook/policy/check-command.policy.ts
